SpaceObject's AddChild/RemoveChild is thread safe now.Unload plugins in reversed order.

+++ World/ISceneManager.h
@@ -12,7 +12,7 @@ namespace Lightning
 			virtual IScene* GetForegroundScene() = 0;
 			virtual void SetForegroundScene(IScene* scene) = 0;
 			virtual void Tick() = 0;
-			virtual void DestroyScene(const std::uint32_t sceneId) = 0;
+			virtual void DestroyScene(const std::uint64_t sceneId) = 0;
 			virtual void DestroyAllScenes() = 0;
 		};
 	}
